backend/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
torch
torchvision
transformers
opencv-python-headless
numpy
pillow
matplotlib
python-dotenv
